⚡ Bolt: Centralize and optimize date parsing/formatting
- Created centralized `src/utils/date.ts` with shared `timestampCache`.
- Optimized `parseIsoDate` to avoid redundant `Date.parse` calls.
- Unified `formatRelativeTime` with support for 'short' and 'long' styles.
- Refactored `GistStore`, `home` route, `GistCard`, and `GistDetail` to use optimized utilities.
- Improved sorting performance from O(N log N) parsing to O(N).
- Reduced memory overhead by consolidating multiple local caches.
