+++ src/components/gist-detail.ts
@@ -9,21 +9,8 @@ import { safeError } from '../services/security/logger';
 import gistStore from '../stores/gist-store';
 import type { GistRecord } from '../types';
 import type { GistRevision } from '../types/api';
+import { formatRelativeTime } from '../utils/date';
 import { toast } from './ui/toast';
-
-function formatRelativeTime(dateStr: string): string {
-  const now = new Date();
-  const date = new Date(dateStr);
-  const diffMs = now.getTime() - date.getTime();
-  const diffSec = Math.floor(diffMs / 1000);
-  if (diffSec < 60) return 'Just now';
-  const diffMin = Math.floor(diffSec / 60);
-  if (diffMin < 60) return `${diffMin}m ago`;
-  const diffHr = Math.floor(diffMin / 60);
-  if (diffHr < 24) return `${diffHr}h ago`;
-  const diffDay = Math.floor(diffHr / 24);
-  return `${diffDay}d ago`;
-}
 
 export function renderFileContent(content: string, language?: string): string {
   return `<pre class="code-block language-${sanitizeHtml(language || 'text')}"><code>${sanitizeHtml(content)}</code></pre>`;
@@ -68,7 +55,7 @@ export function renderGistDetail(gist: GistRecord): string {
           <span class="detail-chip">${fileCount} Files</span>
           <span class="detail-chip">${visibility}</span>
           <time class="micro-label" datetime="${gist.updatedAt}">
-            Updated ${formatRelativeTime(gist.updatedAt)}
+            Updated ${formatRelativeTime(gist.updatedAt, 'long')}
           </time>
         </div>
         <div class="gist-detail-actions">
